feat(core): extract core results-per-page
